subscriber/requirements.txt
pymongo==4.1.1
paho-mqtt==1.6.1
python-dateutil==2.8.2